backend/requirements.txt
Django==4.2
djangorestframework
django-cors-headers
pymongo==4.5.0
mongoengine==0.24.2
bcrypt==4.0.1